backend-api/requirements.txt
fastapi
uvicorn
pydantic-settings
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiosmtplib
httpx
celery
redis
sse-starlette
flower
pytest
pytest-asyncio
